moved gameSetup to features folder
