+++ src/features/gameSetup/PlayerInfoCard.tsx
